Refacter worker related code.

+++ src/thrift/tachyon.thrift
@@ -71,8 +71,8 @@ exception TableDoesNotExistException {
 
 service MasterService {
   // Services to Workers
-  i64 worker_register(1: NetAddress workerNetAddress, 2: i64 totalBytes, 3: i64 usedBytes, 4: list<i64> currentFileList) // Returned value rv % 100,000 is really workerId, rv / 1000,000 is master started time.
-  Command worker_heartbeat(1: i64 workerId, 2: i64 usedBytes, 3: list<i64> removedFileList)
+  i64 worker_register(1: NetAddress workerNetAddress, 2: i64 totalBytes, 3: i64 usedBytes, 4: list<i32> currentFileList) // Returned value rv % 100,000 is really workerId, rv / 1000,000 is master started time.
+  Command worker_heartbeat(1: i64 workerId, 2: i64 usedBytes, 3: list<i32> removedFileList)
   void worker_addFile(1: i64 workerId, 2: i64 workerUsedBytes, 3: i32 fileId, 4: i32 fileSizeBytes, 5: bool hasCheckpointed, 6: string checkpointPath) throws (1: FileDoesNotExistException eP, 2: SuspectedFileSizeException eS)
   set<i32> worker_getPinList()
